feat: clean up and testing
