+++ src/LEGO.AsyncAPI.Readers/AsyncApiReaderSettings.cs
@@ -18,9 +18,14 @@ namespace LEGO.AsyncAPI.Readers
         DoNotResolveReferences,
 
         /// <summary>
-        /// Resolve internal component references and inline them.
+        /// Resolve all references and inline them.
         /// </summary>
-        ResolveReferences,
+        ResolveAllReferences,
+
+        /// <summary>
+        /// Resolve internal component references and inline them while leaving external references as placeholder objects with an AsyncApiReference instance and UnresolvedReference set to true.
+        /// </summary>
+        ResolveInternalReferencesOnly,
     }
 
     /// <summary>
@@ -32,7 +37,7 @@ namespace LEGO.AsyncAPI.Readers
         /// Indicates how references in the source document should be handled.
         /// </summary>
         public ReferenceResolutionSetting ReferenceResolution { get; set; } =
-            ReferenceResolutionSetting.ResolveReferences;
+            ReferenceResolutionSetting.ResolveAllReferences;
 
         /// <summary>
         /// Dictionary of parsers for converting extensions into strongly typed classes.
