+++ test/LEGO.AsyncAPI.Tests/AsyncApiReaderTests.cs
@@ -1,7 +1,5 @@
 // Copyright (c) The LEGO Group. All rights reserved.
 
-using FluentAssertions;
-
 namespace LEGO.AsyncAPI.Tests
 {
     using System;
@@ -592,81 +590,6 @@ namespace LEGO.AsyncAPI.Tests
             Assert.AreEqual(SecuritySchemeType.OAuth2, requirement.Key.Type);
             Assert.IsTrue(requirement.Value.Contains("write:pets"));
             Assert.IsTrue(requirement.Value.Contains("read:pets"));
-        }
-
-        [Test]
-        public void Read_WithExternalResourcesInterfaceDeserializes()
-        {
-          var yaml = """
-                     asyncapi: 2.3.0
-                     info:
-                       title: test
-                       version: 1.0.0
-                     channels:
-                       workspace:
-                         publish:
-                           message:
-                             $ref: "./some/path/to/external/message.yaml"
-                     """;
-          var settings = new AsyncApiReaderSettings
-          {
-            ExternalReferenceReader = new MockExternalReferenceReader(),
-          };
-          var reader = new AsyncApiStringReader(settings);
-          var doc = reader.Read(yaml, out var diagnostic);
-          var message = doc.Channels["workspace"].Publish.Message.First();
-          message.Name.Should().Be("Test");
-          message.Payload.Properties.Count.Should().Be(3);
-        }
-    }
-
-    public class MockExternalReferenceReader : IAsyncApiExternalReferenceReader
-    {
-        public string GetExternalResource(string reference)
-        {
-            if (reference == "./some/path/to/external/message.yaml")
-            {
-              return """
-               name: Test
-               title: Test message
-               summary: Test.
-               schemaFormat: application/schema+yaml;version=draft-07
-               contentType: application/cloudevents+json
-               payload:
-                $ref: "./some/path/to/schema.yaml"
-               """;
-            }
-
-            return """
-                   type: object
-                   properties:
-                     orderId:
-                       description: The ID of the order.
-                       type: string
-                       format: uuid
-                     name:
-                       description: Name of order.
-                       type: string
-                     orderDetails:
-                       description: User details.
-                       type: object
-                       properties:
-                         userId:
-                           description: User Id.
-                           type: string
-                           format: uuid
-                         userName:
-                           description: User name.
-                           type: string
-                   required:
-                   - orderId
-                   example:
-                     orderId: 8f9189f8-653b-4849-a1ec-c838c030bd67
-                     handler: SomeName
-                     orderDetails:
-                       userId: Admin
-                       userName: Admin
-                   """;
         }
     }
 }
